// src/app/models/cong-viec/duyet-ban-giao-cong-viec.ts
export interface TimKiemDanhSachDBGCV {
    nguoiBanGiao?: string;
    nguoiNhanBanGiao?: string;
    noiDungBanGiao?: string;
    tuNgay?: string;
    denngay?: string;
    nguoiNhanId?: number;
    timChinhXac?: number;
}

export interface TimKiemDBGCV {
    keyWord?: string;
    nam?: number;
    thang?: number;
    maHS?: string;
    tieude?: string;
    nguoiLap?: number;
    donVi?: number;
    loaihs?: number;
    phongban?: number;
    trangthai?: number;
    mahs_pb?: string;
    hoso?: string;
    linhvucid?: number;
    timChinhXacSearch?: number;
}